split indexing code into multiple files
