+++ render.go
@@ -7,27 +7,8 @@ import (
 	"io/ioutil"
 	"net/http"
 	"os"
-	"sort"
 	"strings"
 )
-
-func index_file(file os.FileInfo, out chan fileinfo_internal) {
-	var ext string
-	if strings.Contains(file.Name(), ".") {
-		exts := strings.Split(file.Name(), ".")
-		ext = exts[len(exts)-1]
-	} else {
-		ext = ""
-	}
-
-	fileinfo_to_append := fileinfo_internal{
-		Name: file.Name(),
-		Ext:  ext,
-		Dir:  file.IsDir(),
-	}
-
-	out <- fileinfo_to_append
-}
 
 func render(w http.ResponseWriter, r *http.Request) {
 	//slashless_path := config.Path[1:]
@@ -89,48 +70,12 @@ func render(w http.ResponseWriter, r *http.Request) {
 			return
 		}
 
-		files, err := ioutil.ReadDir(dir)
+		fileinfo_arr, err := index_dir(dir)
 		if err != nil {
 			fmt.Println(err)
 			w.Header().Set("Content-Type", "text/plain")
 			io.WriteString(w, err.Error())
 		}
-
-		var fileinfo_arr []fileinfo_internal
-		var fileinfo_dir_arr []fileinfo_internal
-		var fileinfo_file_arr []fileinfo_internal
-
-		file_chan := make(chan fileinfo_internal)
-
-		count := len(files)
-
-		for _, fileinfo := range files {
-			go index_file(fileinfo, file_chan)
-		}
-
-		for count > 0 {
-			fileinfo_arr = append(fileinfo_arr, <-file_chan)
-			count--
-		}
-
-		//Split into two arrays
-		for _, fileinfo := range fileinfo_arr {
-			if fileinfo.Dir {
-				fileinfo_dir_arr = append(fileinfo_dir_arr, fileinfo)
-			} else {
-				fileinfo_file_arr = append(fileinfo_file_arr, fileinfo)
-			}
-		}
-
-		sort.Slice(fileinfo_dir_arr, func(i, j int) bool {
-			return fileinfo_dir_arr[i].Name < fileinfo_dir_arr[j].Name
-		})
-
-		sort.Slice(fileinfo_file_arr, func(i, j int) bool {
-			return fileinfo_file_arr[i].Name < fileinfo_file_arr[j].Name
-		})
-
-		fileinfo_arr = append(fileinfo_dir_arr, fileinfo_file_arr...)
 
 		topmost_dir_arr := strings.Split(pubdir, "/")
 		topmost_dir := topmost_dir_arr[len(topmost_dir_arr)-1]
